+++ src/components/Home/AccountModal/FacebookLoginButton.vue
@@ -1,10 +1,11 @@
 <template>
-    <VFacebookLogin ref="facebook" :app-id="appId" :login-options="loginOptions" @login="login" @sdk-init="initSDK">
+    <VFacebookLogin ref="facebook" :app-id="appId" :login-options="loginOptions" class="btn-block btn-lg"
+                    @login="login" @sdk-init="initSDK">
         <template #login>
             <span v-html="$t('FacebookLoginButton.logInWithFacebook')"/>
         </template>
         <template #working>
-            <Loading :icon-size="15"/>
+            <Loading :icon-size="20"/>
         </template>
     </VFacebookLogin>
 </template>