+++ src/components/Home/AccountModal/LoginPanel.vue
@@ -54,7 +54,17 @@
         </div>
         <div class="row">
             <div class="col-12">
+                <hr class="bg-dark my-1"/>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-12">
                 <FacebookLoginButton class="mt-2 text-center" @hide-account-modal="hideAccountModal"/>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-12 text-center">
+                <GoogleLoginButton class="mt-2"/>
             </div>
         </div>
     </div>
@@ -64,10 +74,11 @@
 import Vue from "vue";
 import RedAsterisk from "../../shared/Forms/RedAsterisk";
 import FacebookLoginButton from "@/components/Home/AccountModal/FacebookLoginButton";
+import GoogleLoginButton from "@/components/Home/AccountModal/GoogleLoginButton";
 
 export default {
     name: "LoginPanel",
-    components: { FacebookLoginButton, RedAsterisk },
+    components: { GoogleLoginButton, FacebookLoginButton, RedAsterisk },
     props: {
         loading: {
             type: Boolean,
@@ -81,9 +92,6 @@ export default {
                 password: undefined,
             },
         };
-    },
-    mounted() {
-        this.focusEmailInput();
     },
     methods: {
         getCredentials() {