+++ src/plugins/vue-gtag/index.js
@@ -2,6 +2,6 @@ import Vue from "vue";
 import VueGtag from "vue-gtag";
 import Config from "../../../config";
 
-if (Config.googleAnalytics.enabled) {
-    Vue.use(VueGtag, { config: { id: Config.googleAnalytics.id } });
+if (Config.google.analytics.enabled) {
+    Vue.use(VueGtag, { config: { id: Config.google.analytics.id } });
 }