+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### 📦 Packages
 
 * `vue-facebook-login-component` installed with version `4.0.1`.
+* `vue-google-login` installed with version `2.0.5`.
 * `@fortawesome/fontawesome-free` updated to version `5.15.3`.
 * `@sentry/browser` updated to version `6.2.3`.
 * `@sentry/integrations` updated to version `6.2.3`.